1,修正销售信息首页 销售信息新增如下几个字段： 2，预计开工时间：能够选择年月，不要日期，放在结束时间字段后面 3，本部门参与人员：多行字段(默认5行)，放在其它部门参与人员前面 4，其它部门参与人员：多行字段(默认5行)，放在最后

+++ modules/200_sales_opportunity_manager_som/app/controllers/som/sales_opportunities_controller.rb
@@ -98,7 +98,7 @@ class Som::SalesOpportunitiesController < ApplicationController
           else
             where_str << "(#{Som::SalesOpportunity.table_name}.possibility > ? and #{Som::SalesOpportunity.table_name}.possibility <= ?)"
           end
-          where_params << p.split("_").collect{|i| i.to_i*10}
+          where_params << p.split("_").collect{|i| i.to_f*10}
         }
 
         sales_opportunities_scope = sales_opportunities_scope.where(([where_str]+where_params).flatten)
